requirements.txt
numpy
pandas
scikit-learn
pytextspan
groupbyrule @ git+https://github.com/OlivierBinette/groupbyrule.git
tqdm
